Add files via upload
This program is used to implement Audio Steganography. It takes ".wav" audio files and ".txt" files as input and merges the text data into the given audio file. The output will be an audio file with hidden text data in it. Audio Steganography is a great way of sending secured messages over public networks where encryption and digital signatures are no longer secured.